normativos-pipeline/requirements.txt
google-generativeai>=0.7.0
supabase>=2.4.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
pypdf2>=3.0.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
tenacity>=8.3.0
rich>=13.7.0
